% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mnist.R
\name{load_mnist}
\alias{load_mnist}
\title{Load MNIST}
\usage{
load_mnist(dirName = "image-data", dat = "train-images.idx3-ubyte",
  labels = "train-labels.idx1-ubyte")
}
\details{
Original data source can be found at \code{http://yann.lecun.com/exdb/mnist/}
}
